refactor utils/config* to config/
