+++ cmd/mattermost/commands/plugin_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/mattermost/mattermost-server/utils"
+	"github.com/mattermost/mattermost-server/config"
 	"github.com/mattermost/mattermost-server/utils/fileutils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -15,11 +15,11 @@ func TestPlugin(t *testing.T) {
 	th := Setup().InitBasic()
 	defer th.TearDown()
 
-	config := th.Config()
-	*config.PluginSettings.EnableUploads = true
-	*config.PluginSettings.Directory = "./test-plugins"
-	*config.PluginSettings.ClientDirectory = "./test-client-plugins"
-	th.SetConfig(config)
+	c := th.Config()
+	*c.PluginSettings.EnableUploads = true
+	*c.PluginSettings.Directory = "./test-plugins"
+	*c.PluginSettings.ClientDirectory = "./test-client-plugins"
+	th.SetConfig(c)
 
 	os.MkdirAll("./test-plugins", os.ModePerm)
 	os.MkdirAll("./test-client-plugins", os.ModePerm)
@@ -31,12 +31,12 @@ func TestPlugin(t *testing.T) {
 	th.CheckCommand(t, "plugin", "add", filepath.Join(path, "testplugin.tar.gz"))
 
 	th.CheckCommand(t, "plugin", "enable", "testplugin")
-	cfg, _, _, err := utils.LoadConfig(th.ConfigPath())
+	cfg, _, _, err := config.LoadConfig(th.ConfigPath())
 	require.Nil(t, err)
 	assert.Equal(t, cfg.PluginSettings.PluginStates["testplugin"].Enable, true)
 
 	th.CheckCommand(t, "plugin", "disable", "testplugin")
-	cfg, _, _, err = utils.LoadConfig(th.ConfigPath())
+	cfg, _, _, err = config.LoadConfig(th.ConfigPath())
 	require.Nil(t, err)
 	assert.Equal(t, cfg.PluginSettings.PluginStates["testplugin"].Enable, false)
